pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "orbit-cli"
version = "0.1.0"
description = "Local-first, multi-model DevOps CLI agent"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.11"
authors = [
    {name = "Abhimanyu Bhagwati"},
]
keywords = ["devops", "cli", "agent", "llm", "ollama", "automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: System :: Systems Administration",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Typing :: Typed",
]
dependencies = [
    "typer>=0.12",
    "rich>=13",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "httpx>=0.27",
    "ollama>=0.4",
    "tomli-w>=1.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
    "mypy>=1.10",
    "types-PyYAML>=6.0",
]
openai = ["openai>=1.0"]
anthropic = ["anthropic>=0.30"]
rag = ["chromadb>=0.5"]
all = ["orbit-cli[dev,openai,anthropic,rag]"]

[project.scripts]
orbit = "orbit.cli:app"

[project.urls]
Homepage = "https://github.com/abhimanyubhagwati/orbit-cli"
Repository = "https://github.com/abhimanyubhagwati/orbit-cli"
Issues = "https://github.com/abhimanyubhagwati/orbit-cli/issues"

[tool.hatch.build.targets.wheel]
packages = ["orbit"]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
